+++ src/service/worker/runtime/graphql/index.ts
@@ -2,19 +2,14 @@ import TokenBucket from 'tokenbucket'
 import { IOClients } from '../../../../clients/IOClients'
 import { nameSpanOperationMiddleware } from '../../../tracing/tracingMiddlewares'
 import { createPrivateHttpRoute } from '../http'
-import {
-  ClientsConfig,
-  GraphQLOptions,
-  ParamsContext,
-  RecorderState,
-  ServiceRoute,
-} from '../typings'
+import { ClientsConfig, GraphQLOptions, ParamsContext, RecorderState, ServiceRoute } from '../typings'
 import { injectGraphqlContext } from './middlewares/context'
 import { graphqlError } from './middlewares/error'
 import { extractQuery } from './middlewares/query'
 import { response } from './middlewares/response'
 import { run } from './middlewares/run'
 import { upload } from './middlewares/upload'
+import { updateSchema } from './middlewares/updateSchema'
 import { makeSchema } from './schema'
 import { GraphQLContext } from './typings'
 
@@ -27,15 +22,16 @@ export const createGraphQLRoute = <T extends IOClients, U extends RecorderState,
   routeId: string,
   globalLimiter: TokenBucket | undefined
 ) => {
-  const schema = makeSchema(graphql)
+  const executableSchema = makeSchema(graphql)
   const pipeline = [
     nameSpanOperationMiddleware('graphql-handler', GRAPHQL_ROUTE),
+    updateSchema(graphql, executableSchema),
     injectGraphqlContext,
     response,
     graphqlError,
     upload,
-    extractQuery(schema),
-    run(schema),
+    extractQuery(executableSchema),
+    run(executableSchema),
   ]
   return createPrivateHttpRoute<T, U, V & GraphQLContext>(clientsConfig, pipeline, serviceRoute, routeId, globalLimiter)
 }