+++ src/service/worker/runtime/graphql/middlewares/run.ts
@@ -1,32 +1,25 @@
-import { execute, GraphQLSchema } from 'graphql'
+import { execute } from 'graphql'
 
-import { GraphQLServiceContext } from '../typings'
+import { GraphQLServiceContext, ExecutableSchema } from '../typings'
 
-export const run = (schema: GraphQLSchema) =>
-async function runHttpQuery (ctx: GraphQLServiceContext, next: () => Promise<void>) {
-  const {
-    graphql: {
-      query,
-    },
-  } = ctx
+export const run = (executableSchema: ExecutableSchema) =>
+  async function runHttpQuery(ctx: GraphQLServiceContext, next: () => Promise<void>) {
+    const {
+      graphql: { query },
+    } = ctx
 
-  const {
-    document,
-    operationName,
-    variables: variableValues,
-  } = query!
+    const { document, operationName, variables: variableValues } = query!
+    const schema = executableSchema.schema
+    const response = await execute({
+      contextValue: ctx,
+      document,
+      fieldResolver: (root, _, __, info) => root[info.fieldName],
+      operationName,
+      rootValue: null,
+      schema,
+      variableValues,
+    })
+    ctx.graphql.graphqlResponse = response
 
-  const response = await execute({
-    contextValue: ctx,
-    document,
-    fieldResolver: (root, _, __, info) => root[info.fieldName],
-    operationName,
-    rootValue: null,
-    schema,
-    variableValues,
-  })
-
-  ctx.graphql.graphqlResponse = response
-
-  await next()
-}
+    await next()
+  }